refactor(suite): move legacy labeling visibility to metadata
fix: split some metadata code to prevent circular imports in metadata package
