+++ suite-common/redux-utils/src/extraDependenciesType.ts
@@ -71,11 +71,9 @@ export type ExtraDependenciesStatic = {
         selectDesktopBinDir: SuiteCompatibleSelector<string | undefined>;
         selectLanguage: SuiteCompatibleSelector<string>;
         selectIsWindowVisible: SuiteCompatibleSelector<boolean>;
-        selectMetadata: SuiteCompatibleSelector<any>;
         selectAddressDisplayType: SuiteCompatibleSelector<AddressDisplayOptions>;
         selectSelectedAccount: SuiteCompatibleSelector<SelectedAccountStatus>;
         selectSelectedAccountStatus: SuiteCompatibleSelector<SelectedAccountStatus['status']>;
-        selectIsSuiteSyncEnabled: SuiteCompatibleSelector<boolean>;
         selectTradingEnvironment: SuiteCompatibleSelector<
             'production' | 'staging' | 'dev' | 'localhost' | undefined
         >;
