+++ suite/metadata/package.json
@@ -16,9 +16,11 @@
         "@reduxjs/toolkit": "2.11.2",
         "@scure/base": "^2.0.0",
         "@suite-common/device": "workspace:*",
+        "@suite-common/message-system": "workspace:*",
         "@suite-common/metadata-types": "workspace:*",
         "@suite-common/redux-utils": "workspace:*",
         "@suite-common/suite-rbf-labels-migrations-types": "workspace:*",
+        "@suite-common/suite-sync": "workspace:*",
         "@suite-common/suite-types": "workspace:*",
         "@suite-common/suite-utils": "workspace:*",
         "@suite-common/test-utils": "workspace:*",
