+++ suite/metadata/src/metadataMiddleware.ts
@@ -1,23 +1,32 @@
-import { type Dispatch } from '@reduxjs/toolkit';
-import { type MiddlewareAPI } from 'redux';
+import { isAnyOf } from '@reduxjs/toolkit';
 
-import { type AnyAction } from '@suite-common/redux-utils';
-import { accountsActions } from '@suite-common/wallet-core';
+import { type AnyAction, createMiddleware } from '@suite-common/redux-utils';
+import { accountsActions, applyDeviceStatesThunk } from '@suite-common/wallet-core';
 
 import { type MetadataAction } from './metadataActions';
 import * as metadataLabelingActions from './metadataLabelingActions';
-import { type MetadataRootState } from './metadataReducer';
+import * as metadataThunks from './metadataThunks';
+import { selectIsLegacyLabelingVisible } from './selectIsLegacyLabelingVisible';
 
 type Action = MetadataAction | ReturnType<typeof accountsActions.createAccount> | AnyAction;
 
-export const metadataMiddleware =
-    (api: MiddlewareAPI<Dispatch, MetadataRootState>) =>
-    (next: Dispatch) =>
-    (action: Action): AnyAction => {
+export const metadataMiddleware = createMiddleware<Action>(
+    (action, { dispatch, getState, next }) => {
         if (accountsActions.createAccount.match(action)) {
-            action.payload = api.dispatch(
+            action.payload = dispatch(
                 metadataLabelingActions.setAccountMetadataKey(action.payload),
             );
+        }
+
+        if (
+            isAnyOf(applyDeviceStatesThunk.fulfilled)(action) &&
+            selectIsLegacyLabelingVisible(getState())
+        ) {
+            const staticSessionId = action.payload.device.state?.staticSessionId;
+
+            if (staticSessionId !== undefined) {
+                dispatch(metadataThunks.initNewDeviceStateMetadataThunk(staticSessionId));
+            }
         }
 
         // pass action
@@ -26,8 +35,8 @@ export const metadataMiddleware =
         switch (action.type) {
             case '@router/location-change': // hack: to prevent dependency
                 // if there is editing field active, changing route turns it inactive
-                if (api.getState().metadata.editing) {
-                    api.dispatch(metadataLabelingActions.setEditing(undefined));
+                if (getState().metadata.editing) {
+                    dispatch(metadataLabelingActions.setEditing(undefined));
                 }
                 break;
             default:
@@ -35,4 +44,5 @@ export const metadataMiddleware =
         }
 
         return action;
-    };
+    },
+);
